feat: move reload logic to router plugin and add new configuration for the logic
